Version du 03/11/2013 - 20131103.001
Cf versions.xml

+++ controllers/plugins_controller.php
@@ -119,19 +119,20 @@ class PluginsController extends AppController {
  * 	Pour ce faire, elle va chercher directement dans la classe du fichier plugin.php les variables qui contiennent les informations nécessaires :
  * 		$plugin->filesCopy			La variable qui contient à la base la liste des fichiers à copier
  *		$plugin->sourceFolder		Détermine le dossier source des fichiers à copier. Le plus souvent, c'est " dirname(__FILE__).DS.'install' ", mais cela peut changer
- *		$plugin->moduleType			Détermine le nom du type de module dans la bdd, celui qui est affiché dans le backoffice
- *		$plugin->tables				Une chaîne ou un tableau contenant la liste des tables dans la BDD qui sont utilisées par le plugin
  *
  * @param 	integer $id Identifiant du plugin à désinstaller
  * @access 	public
  * @author 	koéZionCMS
  * @version 0.1 - 30/03/2013 by AA
+ * @version 0.2 - 29/10/2013 by FI - Reprise de la fonction pour la terminer...
  */
 	function backoffice_uninstall($id) {
 
 		$errors = array();
 		$plugin = $this->Plugin->findFirst(array('conditions' => array('id' => $id)));//On récupère le nom du plugin
-		if ($plugin['installed']) {//On ne réalise la désinstallation que si le plugin est installé
+		
+		//On ne réalise la désinstallation que si le plugin est installé
+		if($plugin['installed']) {
 
 			$pluginName = Inflector::camelize($plugin['code']).'Plugin'; //Génération du nom du plugin			
 			$pluginFile = PLUGINS.DS.$plugin['code'].DS.'plugin.php'; //Chemin vers le fichier d'installation du plugin
@@ -141,73 +142,85 @@ class PluginsController extends AppController {
 				require_once($pluginFile); //Chargement du fichier
 				$pluginClass = new $pluginName(); //Création d'un objet plugin
 
+				//////////////////////////////////
+				//			ACTIONS BDD			//
+				//////////////////////////////////
+												
+				//On va récupérer l'identifiant du type de module à supprimer
+				$modulesTypeIdResult = current($this->Plugin->query("SELECT DISTINCT(`modules_type_id`) FROM `modules` WHERE `plugin_id` = ".$id.";", true));
+				$modulesTypeId = $modulesTypeIdResult['modules_type_id'];				
+				
+				$this->Plugin->query('DELETE FROM `modules` WHERE `plugin_id` = '.$id.';'); //On supprime tous les modules associés à ce plugin
+				$this->Plugin->query('DELETE FROM `plugins` WHERE `id` = '.$id.';'); //On supprime le plugin dans la base de données
+				$this->Plugin->query('DELETE FROM `modules_types` WHERE `id` = '.$modulesTypeId.';'); //On supprime le type de module associé à ce plugin
+								
+				//ON VA RECUPERER, SI IL Y EN A, LA LISTE DES TABLES ASSOCIEES AU PLUGIN EN COURS DE SUPPRESSION
+				$databaseTables = $this->Plugin->table_list_in_database(); //Liste des tables de la BDD
+				$databasePluginTables = array(); //Liste des tables du plugin
+				$databasePluginTablesPrefix = 'plugins_'.$plugin['code'].'_'; //Préfix des tables du plugin
+				
+				//On va parcourir la liste des tables de la base de données pour en extraire les tables associées au plugin
+				//Elles seront ensuite renommées en vue d'une suppression manuelle par le gestionnaire de la BDD
+				foreach($databaseTables as $databaseTable) {
+				
+					if(substr_count($databaseTable, $databasePluginTablesPrefix)) { $sql .= "RENAME TABLE `".$databaseTable."` TO `_".$databaseTable."_".date("Ymd_His")."`;"; }
+				}
+
+				//////////////////////////////////
+				//		ACTIONS FICHIERS		//
+				//////////////////////////////////
+				
 				//Si des fichiers doivent être supprimés, on les récupère dans le plugin
-				if (isset($pluginClass->filesCopy)) {
+				if(isset($pluginClass->filesCopy)) {
 
 					foreach($pluginClass->filesCopy as $fileDelete) {
 
 						if(isset($fileDelete['sourceName']) && isset($fileDelete['destinationName'])) {
-							//Dans le cas ou on un seul fichier à supprimer
 
+							//Dans le cas ou on un seul fichier à supprimer
 							$fileToDelete = $fileDelete['destinationPath'].DS.$fileDelete['destinationName']; //Chemin du fichier de destination
 							$processResult = FileAndDir::remove($fileToDelete); //Suppression du fichier
+							
 						} else if(isset($fileDelete['sourcePath']) && isset($fileDelete['destinationPath'])) {
+							
 							//Dans le cas ou on a le contenu d'un dossier, il faut supprimer tous les fichiers qui ont été créés
-
 							$sourcePathContent = FileAndDir::directoryContent($fileDelete['sourcePath']);
 							foreach($sourcePathContent as $v) {
 
-								$processResult = FileAndDir::remove($fileDelete['destinationPath'].DS.$v);//Suppression du fichier
+								$processResult = FileAndDir::remove($fileDelete['destinationPath'].DS.$v); //Suppression du fichier
 							}
 						}
-						if (isset($fileDelete['removePath']) && $fileDelete['removePath'] == true) {
-							//Si l'index removePath est défini à true, alors on supprime le dossier en plus du fichier
-							FileAndDir::remove_directory($fileDelete['destinationPath']);
-						}
-						if (!$processResult) { $errors[] = $fileDelete['destinationName']; }
+						
+						//Si l'index removePath est défini à true, alors on supprime le dossier en plus du fichier
+						if(isset($fileDelete['removePath']) && $fileDelete['removePath'] == true) { FileAndDir::remove_directory($fileDelete['destinationPath']); }
+						
+						if(!$processResult) { $errors[] = $fileDelete['destinationName']; }
 					}
-					if (!empty($errors)) {
+					if(!empty($errors)) {
+						
 						Session::setFlash('Une erreur est survenue lors de la suppression de certains fichiers :<br />'.implode('<br />* ', $errors), 'error');
 						$this->redirect('backoffice/plugins/index'); //On retourne sur la page de listing
-						return false;
+						
 					}
-				}
-
-				/////////////////////////////////////////////////////////////////
-				////////////// MODIFICATIONS DE LA BASE DE DONNEES //////////////
-				/////////////////////////////////////////////////////////////////
-
-				$this->Plugin->query('DELETE FROM modules WHERE plugin_id = '.$id);//On supprime tous les modules associés à ce plugin
-
-				$this->Plugin->query('DELETE FROM plugins WHERE id = '.$id);//On supprime le plugin dans la base de données
-
-				if (isset($pluginClass->moduleType)) {
-
-					$sql = 'DELETE FROM modules_types WHERE name = "'.$pluginClass->moduleType.'"';//Si un nom de module_type est présent dans le plugin, on le supprime de la base de données
-					$this->Plugin->query($sql);
-				}
-
-				if (isset($pluginClass->tables)) {
-
-					$sql = 'DROP TABLE IF EXISTS '.implode(',', (array) $pluginClass->tables);//Si une liste de tables est présente dans le plugin, on les supprime toutes. On utilise le transtypage pour pouvoir passer une chaîne de caractères OU un tableau
-					$this->Plugin->query($sql);
-				}
+				}				
+				
 			} else {
-				Session::setFlash('Le fichier plugin.php n\'existe pas', 'error');
+				
+				Session::setFlash("Le fichier plugin.php n'existe pas", 'error');
 				$this->redirect('backoffice/plugins/index'); //On retourne sur la page de listing
-				return false;
+								
 			}
 		} else {
+			
 			Session::setFlash('Le plugin doit être installé pour pouvoir être désinstallé', 'error');
 			$this->redirect('backoffice/plugins/index'); //On retourne sur la page de listing
-			return false;
+			
 		}
 		
 		FileAndDir::delete_directory_file(TMP.DS.'cache'.DS.'models'.DS); //Suppression du fichier de cache de la bdd
 		FileAndDir::remove(TMP.DS.'cache'.DS.'variables'.DS.'Plugins'.DS.'plugins.cache'); //Suppression du fichier de cache des plugins
-		Session::setFlash('Le plugin a été correctement désinstallé<br />'.$datas);
+		Session::setFlash('Le plugin a été correctement désinstallé');
 		$this->redirect('backoffice/plugins/index'); //On retourne sur la page de listing
-		return true;
 	}
 	
 /**
